requirements.txt
google-cloud-storage==1.*
rich==10.*